Classes precision changed from float to double. Point3D class added, operations with vectors and matrices added.

+++ VMQMath/src/Quaternion.h
@@ -9,25 +9,25 @@ namespace VMQ {
 
 	public:
 		Quaternion() : m_scalar(0), m_vector() {}
-		Quaternion(float w, float x, float y, float z) :m_scalar(w), m_vector(x, y, z) {}
-		Quaternion(float, const Vector3D&);
-		float W() const;
-		float X() const;
-		float Y() const;
-		float Z() const;
-		float GetScalar() const;
+		Quaternion(double w, double x, double y, double z) :m_scalar(w), m_vector(x, y, z) {}
+		Quaternion(double, const Vector3D&);
+		double W() const;
+		double X() const;
+		double Y() const;
+		double Z() const;
+		double GetScalar() const;
 		const Vector3D& GetVector() const;
-		void SetScalar(float);
+		void SetScalar(double);
 		void SetVector(const Vector3D&);
-		void SetW(float);
-		void SetX(float);
-		void SetY(float);
-		void SetZ(float);
+		void SetW(double);
+		void SetX(double);
+		void SetY(double);
+		void SetZ(double);
 
 		static Quaternion Multiply(const Quaternion&, const Quaternion&);
 
 	private:
-		float m_scalar;
+		double m_scalar;
 		Vector3D m_vector;
 
 	};
